% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WalkingDead2-package.R
\docType{package}
\name{WalkingDead2-package}
\alias{WalkingDead2}
\alias{WalkingDead2-package}
\title{WalkingDead2: Simulate Susceptible-Infected-Immunized State Transitions}
\description{
This package simulates the events of infection (of susceptible subjects) and immunization (of infected subjects) in a population of susceptible, infected, and immunized individuals. The function is still at the development stage and thus has a limited number of arguments that can be altered for the simulation process.
}
\keyword{internal}
