% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WalkingDead2-package.R
\docType{package}
\name{WalkingDead2-package}
\alias{WalkingDead2}
\alias{WalkingDead2-package}
\title{WalkingDead2: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
