% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd2Function.R
\name{wd2}
\alias{wd2}
\alias{plot.wd2}
\alias{summary.wd2}
\title{Walking Dead 2}
\usage{
wd2(
  n,
  m,
  k,
  infect = 0.75,
  infect_radius = 0.075,
  nIterations = 10,
  flight = TRUE
)

\method{plot}{wd2}(x, ...)

\method{summary}{wd2}(object, ...)
}
\arguments{
\item{n, m, k}{Numeric scalars. \code{n} represents the total population size;
\code{m} represents the infected population size; \code{k} represents the
healing population size.}

\item{infect}{numeric value between 0 and 1. This represents the probability
of rate of change from one status to another on contact.}

\item{infect_radius}{numeric value between 0 and 1. Represents the radius within which an
individual can experience a status change when exposed}

\item{nIterations}{positive integer. Represents the number of trials to run
where each individual's status and position may be updated.}

\item{flight}{logical with default set to TRUE. When TRUE, the infected are
programmed to repel from the healers. When FALSE, the infected are programmed
to be attracted to the non-infected.
infected.}

\item{x}{an object of class \code{wd2}}

\item{...}{additional arguments affecting the summary produced.}

\item{object}{an object of class \code{wd2}}
}
\description{
This package simulates the events of infection and healing
in a population of susceptible individuals, infected individuals, and healed
individuals. The function is only at the development stage and thus has a
limited number of arguments that can be altered for the simulation process.
}
\examples{
require(data.table)
temp <- wd2(n=100, m=15, k=5)
temp2 <- summary(temp)
temp3 <- plot(temp)
# head(temp2)
# temp3
}
